+++ gulpfile.js
@@ -1,0 +1,3 @@
+import { dev, build } from './gulp/index.js';
+
+export { dev as default, dev, build }