+++ gulp/config/base.config.js
@@ -1,0 +1,6 @@
+export const baseConfig = {
+  sourceFolder: './src/assets',
+  buildFolder: './assets',
+  isProxy: false,
+  localWebsiteURL: 'http://localhost'
+};