+++ gulp/config/custom.config.js
@@ -1,0 +1,7 @@
+import { baseConfig } from './base.config.js';
+
+export default {
+  ...baseConfig,
+  // rewrite your config here
+  buildFolder: './dist',
+};