+++ gulp/config/rme-theme.config.js
@@ -1,0 +1,8 @@
+import { baseConfig } from './base.config.js';
+
+export default {
+  ...baseConfig,
+  sourceFolder: './src/rem-markup-engine-theme',
+  isProxy: true,
+  localWebsiteURL: 'http://my-domain.loc'
+};