+++ gulp/globals/path.js
@@ -3,15 +3,15 @@ const { default: config } = await import(`../config/${configName}.config.js`);
 
 const paths = {
   source: {
-    scss: `${config.sourceFolder}/**/scss/*.scss`,
-    allScssComponents: `${config.sourceFolder}/**/scss/**/*.scss`,
+    scss: `${config.sourceFolder}/scss/*.scss`,
+    allScssComponents: `${config.sourceFolder}/scss/**/*.scss`,
     js: `${config.sourceFolder}'/{js,**/js}/*.js'`,
     jsEsBuild: `${config.sourceFolder}/js/`,
-    jsLibraries: `${config.sourceFolder}/**/js/libs/*.*`,
-    images: `${config.sourceFolder}/**/img/**/*.{jpg,jpeg,png,gif,webp}`,
-    svgs: `${config.sourceFolder}/**/img/**/*.svg`,
-    fonts: `${config.sourceFolder}/**/fonts/**/*.*`,
-    video: `${config.sourceFolder}/**/videos/**/*.*`,
+    jsLibraries: `${config.sourceFolder}/js/libs/*.*`,
+    images: `${config.sourceFolder}/img/**/*.{jpg,jpeg,png,gif,webp}`,
+    svgs: `${config.sourceFolder}/img/**/*.svg`,
+    fonts: `${config.sourceFolder}/fonts/**/*.*`,
+    video: `${config.sourceFolder}/videos/**/*.*`,
     php: '*.php',
     html: `*.{html, htm}`
   },