+++ CHANGELOG.md
@@ -10,11 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 - New RME theme submodule directory: `src/assets/theme` → `src/rem-markup-engine-theme`
 - Example source files replaces with `.gitkeep` to save folder structure
-- Gulp entry point (**breaking changes**)
+- Gulp entry point
 
 ### Fixed
 
 - Scripts watch process
+
+### Removed
+
+- Unused variables in `glp/globals/paths.js`
 
 ## [0.1-beta.1] — 2026-02-05