+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - New RME theme submodule directory: `src/assets/theme` → `src/rem-markup-engine-theme`
 - Example source files replaces with `.gitkeep` to save folder structure
 - Gulp entry point
+- Path resolution variables for `browser-sync`
 
 ### Fixed