+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - New RME theme submodule directory: `src/assets/theme` → `src/rem-markup-engine-theme`
 - Example source files replaces with `.gitkeep` to save folder structure
 - Gulp entry point (`gulpfile.js` is now a file and gulp folder renamed to `gulp`)
-- Path resolution variables for `browser-sync`
+- Path resolution variables for source assets and `browser-sync`
 
 ### Fixed